Finished ArrayList of pieces construction
Made a buildPieces() that creates an arrayList of Pieces. This list is created based on the subclasses that exist of Pieces and the amount specified in each subclass. So the pieces are constructed directly based on the values set in the subclass. Therefore any set of subclasses that extend Piece are automatically added to the ArrayList based on the amount set in the subclass of Piece.