Final Commit
Finalised the setup from the backend. Currently used a test setup, so that towards the front end there are some pieces. Currently used @Depricated code, because of the test setup. But for a while this should actually be useful, until we have a proper setup. Currently prints the setup, so we can check the backend with the frontend easily.